+++ core/src/main/java/arez/Observer.java
@@ -11,7 +11,6 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import org.realityforge.anodoc.TestOnly;
 import static org.realityforge.braincheck.Guards.*;
 
 /**
@@ -868,7 +867,6 @@ public final class Observer
     return _component;
   }
 
-  @TestOnly
   void markAsScheduled()
   {
     _scheduled = true;